# src/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Show My Your Space",
    "description": "당신의 방을 자랑하세요",
    "version": "1.0.0"
  },
  "host": "localhost:3000/api",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/user/signup": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "회원가입 페이지",
        "description": "회원가입 페이지",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "useremail": {
                  "example": "any"
                },
                "nickname": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "checkpassword": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/user/signin": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "로그인 페이지",
        "description": "로그인 페이지",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "useremail": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/user/signin/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "유저 토큰 확인 페이지",
        "description": "유저 토큰 확인 페이지",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/board": {
      "post": {
        "tags": [
          "Board"
        ],
        "summary": "게시글 작성 페이지",
        "description": "게시글 작성 페이지",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "category": {
                  "example": "any"
                },
                "title": {
                  "example": "any"
                },
                "content": {
                  "example": "any"
                },
                "image1": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "tags": [
          "Board"
        ],
        "summary": "게시글 전체 조회 페이지",
        "description": "게시글 전체 조회 페이지",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/board/{boardId}": {
      "get": {
        "tags": [
          "Board"
        ],
        "summary": "게시글 상세 조회 페이지",
        "description": "게시글 상세 조회 페이지",
        "parameters": [
          {
            "name": "boardId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/boardcate/{category}": {
      "get": {
        "tags": [
          "Board"
        ],
        "summary": "카테고리별 게시글 조회 페이지",
        "description": "카테고리별 게시글 조회 페이지",
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/board/{boardId}/update/": {
      "put": {
        "tags": [
          "Board"
        ],
        "summary": "게시글 수정 페이지",
        "description": "게시글 수정 페이지",
        "parameters": [
          {
            "name": "boardId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "category": {
                  "example": "any"
                },
                "title": {
                  "example": "any"
                },
                "content": {
                  "example": "any"
                },
                "image1": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/board/{boardId}/delete/": {
      "delete": {
        "tags": [
          "Board"
        ],
        "summary": "게시글 삭제 페이지",
        "description": "게시글 삭제 페이지",
        "parameters": [
          {
            "name": "boardId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/": {
      "get": {
        "tags": [
          "Comment"
        ],
        "summary": "댓글 조회 페이지",
        "description": "댓글 조회 페이지",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Comment"
        ],
        "summary": "댓글 작성 페이지",
        "description": "댓글 작성 페이지",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "comment": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/{commentId}": {
      "put": {
        "tags": [
          "Comment"
        ],
        "summary": "댓글 수정 페이지",
        "description": "댓글 수정 페이지",
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "comment": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Comment"
        ],
        "summary": "댓글 삭제 페이지",
        "description": "댓글 삭제 페이지",
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/like/{target}/{Id}": {
      "put": {
        "tags": [
          "Like"
        ],
        "summary": "좋아요 누르기 (게시글, 댓글)",
        "description": "좋아요 누르기 (게시글, 댓글)",
        "parameters": [
          {
            "name": "target",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/cancellike/{target}/{Id}": {
      "put": {
        "tags": [
          "Like"
        ],
        "summary": "좋아요 취소 (게시글, 댓글)",
        "description": "좋아요 취소 (게시글, 댓글)",
        "parameters": [
          {
            "name": "target",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  }
}